third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-fetch-on-dedicated-worker-blocked.https.html

<!DOCTYPE html>
<p>This test starts a dedicated worker that loads insecure data via Fetch API.
We should trigger a mixed content callback and block the load.</p>
<p>The test passes if the result is "LOAD FAILED".</p>
<div id=result></div>
<script>
if (window.testRunner) {
  testRunner.waitUntilDone();
  testRunner.dumpAsText();
  testRunner.overridePreference("WebKitAllowRunningInsecureContent", false);
}

function log(message) {
  document.getElementById("result").innerHTML += message + "<br>";
}

var worker = new Worker("./resources/insecure-fetch-on-dedicated-worker.js");
worker.addEventListener("message", function (e) {
  log("Result: " + e.data);
  if (window.testRunner)
    testRunner.notifyDone();
}, false);
</script>
