third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-fetch-on-dedicated-worker-blocked.https-expected.txt

CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/insecure-fetch-on-dedicated-worker-blocked.https.html' was loaded over HTTPS, but requested an insecure resource 'http://example.test:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*'. This request has been blocked; the content must be served over HTTPS.
This test starts a dedicated worker that loads insecure data via Fetch API. We should trigger a mixed content callback and block the load.

The test passes if the result is "LOAD FAILED".

Result: LOAD FAILED
